MAINT: Move utility block index parser from old thm_model
